Split code into files and add some documentation (and notes).
